# openapi/internal-ai-v1.json
{
  "components": {
    "schemas": {
...
      "ErrorDetail": {
        "additionalProperties": false,
        "properties": {
          "code": {
            "title": "Code",
            "type": "string"
          },
          "message": {
            "title": "Message",
            "type": "string"
          }
        },
        "required": [
          "code",
          "message"
        ],
        "title": "ErrorDetail",
        "type": "object"
      },
      "ErrorResponse": {
        "additionalProperties": false,
        "properties": {
          "error": {
            "$ref": "#/components/schemas/ErrorDetail"
          }
        },
        "required": [
          "error"
        ],
        "title": "ErrorResponse",
        "type": "object"
      },
      "EvidenceReference": {
        "additionalProperties": false,
        "properties": {
          "patientId": {
            "format": "uuid",
            "title": "Patientid",
            "type": "string"
          },
          "sourceId": {
            "format": "uuid",
            "title": "Sourceid",
            "type": "string"
          },
          "sourceType": {
            "enum": [
              "TIMELINE_EVENT",
              "TASK"
            ],
            "title": "Sourcetype",
            "type": "string"
          }
        },
        "required": [
          "sourceType",
          "sourceId",
          "patientId"
        ],
        "title": "EvidenceReference",
        "type": "object"
      },
...
      "GenerateHandoffRequest": {
        "additionalProperties": false,
        "properties": {
          "includeUnverified": {
            "title": "Includeunverified",
            "type": "boolean"
          },
          "patients": {
            "items": {
              "$ref": "#/components/schemas/HandoffPatientSnapshot"
            },
            "maxItems": 1000,
            "title": "Patients",
            "type": "array"
          },
          "precheckItems": {
            "items": {
              "$ref": "#/components/schemas/HandoffPrecheckItem"
            },
            "maxItems": 1000,
            "title": "Precheckitems",
            "type": "array"
          },
          "requestId": {
            "format": "uuid",
            "title": "Requestid",
            "type": "string"
          },
          "templateId": {
            "const": "NURSING_HANDOFF_V1",
            "title": "Templateid",
            "type": "string"
          }
        },
        "required": [
          "requestId",
          "templateId",
          "includeUnverified",
          "patients",
          "precheckItems"
        ],
        "title": "GenerateHandoffRequest",
        "type": "object"
      },
      "GenerateHandoffResponse": {
        "additionalProperties": false,
        "properties": {
          "contractVersion": {
            "pattern": "^[A-Za-z0-9][A-Za-z0-9._:-]{0,63}$",
            "title": "Contractversion",
            "type": "string"
          },
          "generatedAt": {
            "format": "date-time",
            "title": "Generatedat",
            "type": "string"
          },
          "modelVersion": {
            "pattern": "^[A-Za-z0-9][A-Za-z0-9._:-]{0,63}$",
            "title": "Modelversion",
            "type": "string"
          },
          "patients": {
            "items": {
              "$ref": "#/components/schemas/HandoffPatientDraft"
            },
            "maxItems": 1000,
            "title": "Patients",
            "type": "array"
          },
          "requestId": {
            "format": "uuid",
            "title": "Requestid",
            "type": "string"
          },
          "warnings": {
            "items": {
              "$ref": "#/components/schemas/HandoffWarning"
            },
            "maxItems": 1000,
            "title": "Warnings",
            "type": "array"
          }
        },
        "required": [
          "requestId",
          "modelVersion",
          "contractVersion",
          "generatedAt",
          "patients",
          "warnings"
        ],
        "title": "GenerateHandoffResponse",
        "type": "object"
      },
      "HandoffPatientDraft": {
        "additionalProperties": false,
        "properties": {
          "patientId": {
            "format": "uuid",
            "title": "Patientid",
            "type": "string"
          },
          "sections": {
            "items": {
              "$ref": "#/components/schemas/HandoffSection"
            },
            "maxItems": 6,
            "minItems": 6,
            "title": "Sections",
            "type": "array"
          }
        },
        "required": [
          "patientId",
          "sections"
        ],
        "title": "HandoffPatientDraft",
        "type": "object"
      },
      "HandoffPatientSnapshot": {
        "additionalProperties": false,
        "properties": {
          "patientId": {
            "format": "uuid",
            "title": "Patientid",
            "type": "string"
          },
          "tasks": {
            "items": {
              "$ref": "#/components/schemas/SnapshotTask"
            },
            "maxItems": 1000,
            "title": "Tasks",
            "type": "array"
          },
          "timelineEvents": {
            "items": {
              "$ref": "#/components/schemas/TimelineEvent"
            },
            "maxItems": 1000,
            "title": "Timelineevents",
            "type": "array"
          }
        },
        "required": [
          "patientId",
          "timelineEvents",
          "tasks"
        ],
        "title": "HandoffPatientSnapshot",
        "type": "object"
      },
      "HandoffPrecheckItem": {
        "additionalProperties": false,
        "properties": {
          "answer": {
            "anyOf": [
              {
                "enum": [
                  "NO_ISSUE",
                  "INCLUDE_HANDOFF",
                  "UNVERIFIED",
                  "NOT_APPLICABLE"
                ],
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Answer"
          },
          "evidence": {
            "items": {
              "$ref": "#/components/schemas/EvidenceReference"
            },
            "maxItems": 1000,
            "minItems": 1,
            "title": "Evidence",
            "type": "array"
          },
          "id": {
            "format": "uuid",
            "title": "Id",
            "type": "string"
          },
          "question": {
            "maxLength": 1000,
            "minLength": 1,
            "pattern": "^[^\\x00-\\x1F\\x7F]*$",
            "title": "Question",
            "type": "string"
          },
          "severity": {
            "enum": [
              "CRITICAL",
              "RECOMMENDED"
            ],
            "title": "Severity",
            "type": "string"
          }
        },
        "required": [
          "id",
          "severity",
          "question",
          "answer",
          "evidence"
        ],
        "title": "HandoffPrecheckItem",
        "type": "object"
      },
      "HandoffPrecheckRequest": {
        "additionalProperties": false,
        "properties": {
          "patients": {
            "items": {
              "$ref": "#/components/schemas/HandoffPatientSnapshot"
            },
            "maxItems": 1000,
            "title": "Patients",
            "type": "array"
          },
          "requestId": {
            "format": "uuid",
            "title": "Requestid",
            "type": "string"
          }
        },
        "required": [
          "requestId",
          "patients"
        ],
        "title": "HandoffPrecheckRequest",
        "type": "object"
      },
      "HandoffPrecheckResponse": {
        "additionalProperties": false,
        "properties": {
          "contractVersion": {
            "pattern": "^[A-Za-z0-9][A-Za-z0-9._:-]{0,63}$",
            "title": "Contractversion",
            "type": "string"
          },
          "generatedAt": {
            "format": "date-time",
            "title": "Generatedat",
            "type": "string"
          },
          "modelVersion": {
            "pattern": "^[A-Za-z0-9][A-Za-z0-9._:-]{0,63}$",
            "title": "Modelversion",
            "type": "string"
          },
          "questions": {
            "items": {
              "$ref": "#/components/schemas/HandoffQuestion"
            },
            "maxItems": 1000,
            "title": "Questions",
            "type": "array"
          },
          "requestId": {
            "format": "uuid",
            "title": "Requestid",
            "type": "string"
          }
        },
        "required": [
          "requestId",
          "modelVersion",
          "contractVersion",
          "generatedAt",
          "questions"
        ],
        "title": "HandoffPrecheckResponse",
        "type": "object"
      },
      "HandoffQuestion": {
        "additionalProperties": false,
        "properties": {
          "evidence": {
            "items": {
              "$ref": "#/components/schemas/EvidenceReference"
            },
            "maxItems": 1000,
            "minItems": 1,
            "title": "Evidence",
            "type": "array"
          },
          "patientId": {
            "format": "uuid",
            "title": "Patientid",
            "type": "string"
          },
          "prompt": {
            "maxLength": 1000,
            "minLength": 1,
            "pattern": "^[^\\x00-\\x1F\\x7F]*$",
            "title": "Prompt",
            "type": "string"
          },
          "questionKey": {
            "maxLength": 128,
            "minLength": 1,
            "title": "Questionkey",
            "type": "string"
          },
          "reason": {
            "maxLength": 1000,
            "minLength": 1,
            "pattern": "^[^\\x00-\\x1F\\x7F]*$",
            "title": "Reason",
            "type": "string"
          },
          "severity": {
            "enum": [
              "CRITICAL",
              "RECOMMENDED"
            ],
            "title": "Severity",
            "type": "string"
          }
        },
        "required": [
          "questionKey",
          "patientId",
          "severity",
          "prompt",
          "reason",
          "evidence"
        ],
        "title": "HandoffQuestion",
        "type": "object"
      },
      "HandoffSection": {
        "additionalProperties": false,
        "properties": {
          "citations": {
            "items": {
              "$ref": "#/components/schemas/EvidenceReference"
            },
            "maxItems": 1000,
            "title": "Citations",
            "type": "array"
          },
          "content": {
            "maxLength": 5000,
            "minLength": 1,
            "pattern": "^[^\\x00-\\x1F\\x7F]*$",
            "title": "Content",
            "type": "string"
          },
          "section": {
            "enum": [
              "PATIENT_STATUS",
              "PAIN",
              "TREATMENT",
              "DIET",
              "ACTIVITY",
              "OBSERVATION"
            ],
            "title": "Section",
            "type": "string"
          }
        },
        "required": [
          "section",
          "content",
          "citations"
        ],
        "title": "HandoffSection",
        "type": "object"
      },
      "HandoffWarning": {
        "additionalProperties": false,
        "properties": {
          "code": {
            "const": "UNVERIFIED_INFORMATION",
            "title": "Code",
            "type": "string"
          },
          "evidence": {
            "items": {
              "$ref": "#/components/schemas/EvidenceReference"
            },
            "maxItems": 1000,
            "minItems": 1,
            "title": "Evidence",
            "type": "array"
          },
          "itemId": {
            "format": "uuid",
            "title": "Itemid",
            "type": "string"
          },
          "message": {
            "maxLength": 1000,
            "minLength": 1,
            "pattern": "^[^\\x00-\\x1F\\x7F]*$",
            "title": "Message",
            "type": "string"
          },
          "patientId": {
            "format": "uuid",
            "title": "Patientid",
            "type": "string"
          }
        },
        "required": [
          "code",
          "itemId",
          "patientId",
          "message",
          "evidence"
        ],
        "title": "HandoffWarning",
        "type": "object"
      },
...
      "SnapshotTask": {
        "additionalProperties": false,
        "properties": {
          "dueAt": {
            "anyOf": [
              {
                "format": "date-time",
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Dueat"
          },
          "effectivePriority": {
            "enum": [
              "CRITICAL",
              "HIGH",
              "NORMAL"
            ],
            "title": "Effectivepriority",
            "type": "string"
          },
          "id": {
            "format": "uuid",
            "title": "Id",
            "type": "string"
          },
          "sourceReferences": {
            "items": {
              "maxLength": 1000,
              "minLength": 1,
              "pattern": "^[^\\x00-\\x1F\\x7F]*$",
              "type": "string"
            },
            "maxItems": 1000,
            "title": "Sourcereferences",
            "type": "array"
          },
          "title": {
            "maxLength": 200,
            "minLength": 1,
            "pattern": "^[^\\x00-\\x1F\\x7F]*$",
            "title": "Title",
            "type": "string"
          },
          "version": {
            "minimum": 0.0,
            "title": "Version",
            "type": "integer"
          }
        },
        "required": [
          "id",
          "title",
          "dueAt",
          "effectivePriority",
          "version",
          "sourceReferences"
        ],
        "title": "SnapshotTask",
        "type": "object"
      },
...
      "TimelineEvent": {
        "additionalProperties": false,
        "properties": {
          "id": {
            "format": "uuid",
            "title": "Id",
            "type": "string"
          },
          "occurredAt": {
            "format": "date-time",
            "title": "Occurredat",
            "type": "string"
          },
          "sourceReference": {
            "maxLength": 1000,
            "minLength": 1,
            "pattern": "^[^\\x00-\\x1F\\x7F]*$",
            "title": "Sourcereference",
            "type": "string"
          },
          "summary": {
            "maxLength": 1000,
            "minLength": 1,
            "pattern": "^[^\\x00-\\x1F\\x7F]*$",
            "title": "Summary",
            "type": "string"
          },
          "type": {
            "enum": [
              "OBSERVATION",
              "MEDICATION",
              "PROCEDURE",
              "REPORT",
              "TASK"
            ],
            "title": "Type",
            "type": "string"
          }
        },
        "required": [
          "id",
          "occurredAt",
          "type",
          "summary",
          "sourceReference"
        ],
        "title": "TimelineEvent",
        "type": "object"
      }
    },
    "securitySchemes": {
      "APIKeyHeader": {
        "in": "header",
        "name": "X-Internal-Token",
        "type": "apiKey"
      }
    }
  },
  "info": {
    "title": "Nurse Hand Internal AI API",
    "version": "1.0.0"
  },
  "openapi": "3.1.0",
  "paths": {
...
    "/internal/v1/handoffs/generate": {
      "post": {
        "operationId": "generate_handoff_internal_v1_handoffs_generate_post",
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/GenerateHandoffRequest"
              }
            }
          },
          "required": true
        },
        "responses": {
          "200": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/GenerateHandoffResponse"
                }
              }
            },
            "description": "Successful Response"
          },
          "401": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorResponse"
                }
              }
            },
            "description": "Unauthorized"
          },
          "422": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorResponse"
                }
              }
            },
            "description": "Unprocessable Entity"
          },
          "429": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorResponse"
                }
              }
            },
            "description": "Too Many Requests"
          },
          "502": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorResponse"
                }
              }
            },
            "description": "Bad Gateway"
          },
          "503": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorResponse"
                }
              }
            },
            "description": "Service Unavailable"
          },
          "504": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorResponse"
                }
              }
            },
            "description": "Gateway Timeout"
          }
        },
        "security": [
          {
            "APIKeyHeader": []
          }
        ],
        "summary": "Generate Handoff",
        "tags": [
          "internal-handoffs"
        ]
      }
    },
    "/internal/v1/handoffs/precheck": {
      "post": {
        "operationId": "precheck_handoff_internal_v1_handoffs_precheck_post",
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/HandoffPrecheckRequest"
              }
            }
          },
          "required": true
        },
        "responses": {
          "200": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HandoffPrecheckResponse"
                }
              }
            },
            "description": "Successful Response"
          },
          "401": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorResponse"
                }
              }
            },
            "description": "Unauthorized"
          },
          "422": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorResponse"
                }
              }
            },
            "description": "Unprocessable Entity"
          },
          "429": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorResponse"
                }
              }
            },
            "description": "Too Many Requests"
          },
          "502": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorResponse"
                }
              }
            },
            "description": "Bad Gateway"
          },
          "503": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorResponse"
                }
              }
            },
            "description": "Service Unavailable"
          },
          "504": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorResponse"
                }
              }
            },
            "description": "Gateway Timeout"
          }
        },
        "security": [
          {
            "APIKeyHeader": []
          }
        ],
        "summary": "Precheck Handoff",
        "tags": [
          "internal-handoffs"
        ]
      }
    },
...
  }
}